+++ src/Subscriber/CategoryPageSubscriber.php
@@ -11,19 +11,14 @@ use Omikron\FactFinder\Shopware6\Utilites\Ssr\Field\CategoryPath;
 use Shopware\Core\Content\Category\SalesChannel\AbstractCategoryRoute;
 use Shopware\Storefront\Page\Navigation\NavigationPageLoadedEvent;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
-
 use function Omikron\FactFinder\Shopware6\Internal\Utils\safeGetByName;
 
 class CategoryPageSubscriber implements EventSubscriberInterface
 {
     private AbstractCategoryRoute $cmsPageRoute;
-
     private Communication $config;
-
     private ExtensionConfig $extensionConfig;
-
     private string $fieldName;
-
     private array $addParams;
 
     public function __construct(
@@ -68,21 +63,38 @@ class CategoryPageSubscriber implements EventSubscriberInterface
 
         $categoryPath  = (new CategoryPath($this->fieldName))->getValue($category);
         $communication = [
-            'add-params'       => implode(',', array_map(fn (string $key, string $value): string => sprintf('%s=%s', $key, $value), array_keys($mergedAddParams), array_values($mergedAddParams))),
-        ] + ($isCategory ? ['category-page' => $categoryPath] : []);
+                'add-params'       => implode(',', array_map(fn (string $key, string $value): string => sprintf('%s=%s', $key, $value), array_keys($mergedAddParams), array_values($mergedAddParams))),
+            ] + ($isCategory ? ['category-page' => $categoryPath] : []);
 
         if ($route === 'frontend.navigation.page') {
             $event->getRequest()->attributes->set('categoryPath', $categoryPath);
         }
 
+        $factfinderConfig    = $event->getPage()->getExtension('factfinder')->getVars();
+        $communicationConfig = [];
+
+        if (isset($factfinderConfig['communication'])) {
+            $communicationConfig = array_merge($factfinderConfig['communication'], $communication);
+        }
+
         $event->getPage()->getExtension('factfinder')->assign(
             [
-                'communication'         => $communication,
-                'trackingSettings'      => $this->extensionConfig->getTrackingSettings(),
-                'redirectMapping'       => (string) $this->extensionConfig->getRedirectMapping(),
-                'searchImmediate'       => $searchImmediate ? 'true' : 'false',
-                'categoryPathFieldName' => "{$this->fieldName}ROOT",
+                'communication'           => $communication,
+                'trackingSettings'        => $this->extensionConfig->getTrackingSettings(),
+                'redirectMapping'         => (string) $this->extensionConfig->getRedirectMapping(),
+                'searchImmediate'         => $searchImmediate ? 'true' : 'false',
+                'categoryPathFieldName'   => "{$this->fieldName}ROOT",
+                'communicationAttributes' => !empty($communicationConfig) ? $this->getCommunicationAttributes($communicationConfig) : $communication,
             ]
+        );
+    }
+
+    private function getCommunicationAttributes(array $communicationConfig): array
+    {
+        return array_map(
+            fn (string $key, string $value) => sprintf('%s="%s" ', $key, htmlspecialchars($value)),
+            array_keys($communicationConfig),
+            array_values($communicationConfig)
         );
     }
 }