+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## Unreleased
+### Fix
+  - Handle category-page for ff-communication
 ## [v2.0.0] - 2024.05.08
 ### BREAKING
 - IMPORTANT! Drop Shopware 6.5 compatibility