+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ### Fix
   - Handle category-page for ff-communication
   - Fix problem with offcanvas-filter plugin
-  - 
+  - Fix SSR for category page
+
 ## [v2.0.0] - 2024.05.08
 ### BREAKING
 - IMPORTANT! Drop Shopware 6.5 compatibility