+++ CHANGELOG.md
@@ -4,10 +4,10 @@
 - Add userId for cart tracking
 ### Change
 - Support only Shopware v6.5.*
-
 ### Fix
 - Support sid for SSR
 - Fix problem with session data for login tracking
+- Handle category-page attribute for ff-communication
 
 ## [v5.1.0] - 2024.03.11
 ### Add